Fix linting issues in detection.py, move `COCOmAP` to `segmetrics.deprecated`
