+++ requirements.txt
@@ -3,3 +3,4 @@ scipy
 scikit-image>=0.18
 scikit-learn
 dill
+Deprecated==1.2
